Move HapticInterfaceManager to managers.video package
